routing button merge
