+++ app/index.js
@@ -83,16 +83,12 @@ export default function Login() {
 						<Button
 							width="w-[85%]"
 							title="Login"
+							href="/feed"
 							press={() => {
 								console.log("click");
 								/* Keyboard.dismiss();
 								userLogin(email, pass); */
 							}}
-						/>
-						<RoutingButton
-							title="login"
-							width="w-[85%]"
-							href="/feed"
 						/>
 						<StyledText className="text-offwhite text-center text-[18px] mt-5">
 							Don't have an account?{" "}
